requirements.txt
fastapi
uvicorn
transformers
pillow
pydantic
torch
python-multipart
streamlit